Clean up
- Modified Tree::search() to also print and count access to the leaf node.
- Modified Tree::displaySingleNode() to print node keys wrapped in curly braces
- Modified Disk::printBlock() to print only the 'tconst' attribute of records
- Updated logic in Disk::getBlockId
- Cleaned up main.cpp for a cleaner output, mainly up to experiment 3 only.

+++ src/disk.h
@@ -31,7 +31,7 @@ public:
 
     void printRecord(Record *record);
 
-    int getBlockId(Record *record);
+    size_t getBlockId(Record *record);
 
     void printBlock(size_t aBlockIdx);
